avcodec/itut35: always check the provider code and country code together
ITU-T T.35 provider codes are attributed by national bodies and it's
possible to have collisions across countries. This is why the country code
must always be checked as well.

Use if statements rather than nested switches which would be unreadable.

Rename some of the constants to match the corresponding organization.
Add a constant for AOM.
Write all constants with 4 hex digits to make it clear that they are 2-byte ids.

The code for V-Nova should be 0x5000 instead of 0x0050 according to the
UK Register of Manufacturer Codes https://www.cix.co.uk/~bpechey/H221/h221code.htm
but I have been unable to find any reference for what code should be used for
LCEVC, and the author of the original change has not replied to my emails.

Signed-off-by: Maryla Ustarroz-Calonge <maryla@google.com>